Discovered that the issue with the C++ communication example receiving double-firmatified echoes is not due to a problem with the C++ solution, the problem lies with the Python solution and the Arduino program made to support it.
It turns out that unlike the openFrameworks and Arudino implementations of Firmata, when pyfirmata2 sends out SysEx messages it does not automatically convert the payload stream from 8 bits to 7-bit messages. This needs to be done by the caller.

Since we saw no issues with the Arduino interpreting the received data, we assumed that the Arduino was automatically de-firmatifying it. This is not the case, and that is why when echoed back to the C++ solution the received echo was double firmatified.

Therefore, the Python solution needs to be changed to firmatify data before calling send_sysex.

This also supports a change in data processing philosophy. Both Arduino and openFrameworks Firmata implementations do not allow for pre-firmatified data to be sent through sendSysex (or equivalent method), as *any* data passed to sendSysex will be firmatified. We should replicate this procedure when using pyfirmata2, meaning all data received on any device is firmatified. Since we know all valid received data is firmatified, we can defirmatify in the SysEx handlers of each program. While this is unwieldy with pyfirmata2, it should be applied to all other programs.